Email.
Moved checkGrades to Mailer. Added loop that will email each trainer in
the list of trainers who need to submit grades. It's commented out for
now, so we can keep using the test email account. Need to add pretty
formatting to the email next.
