+++ caliber/src/main/java/com/revature/caliber/services/EmailService.java
@@ -37,21 +37,6 @@ public class EmailService {
 	@Autowired
 	private Mailer mailer;
 
-	@Autowired
-	private AssessmentDAO assessmentDAO;
-
-	@Autowired
-	private BatchDAO batchDAO;
-
-	@Autowired
-	private TraineeDAO traineeDAO;
-
-	@Autowired
-	private GradeDAO gradeDAO;
-	
-	@Autowired
-	private TrainerDAO trainerDAO;
-
 	private static final long DAYS_IN_WEEK = 7;
 	private static final int YEAR = 2017;
 	private static final int MONTH = 9;
@@ -62,26 +47,6 @@ public class EmailService {
 
 	public void setMailer(Mailer mailer) {
 		this.mailer = mailer;
-	}
-
-	public void setGrade(GradeDAO gradeDAO) {
-		this.gradeDAO = gradeDAO;
-	}
-
-	public void setAssessmentDAO(AssessmentDAO assessmentDAO) {
-		this.assessmentDAO = assessmentDAO;
-	}
-
-	public void setBatch(BatchDAO batchDAO) {
-		this.batchDAO = batchDAO;
-	}
-
-	public void setTrainee(TraineeDAO traineeDAO) {
-		this.traineeDAO = traineeDAO;
-	}
-	
-	public void setTrainer(TrainerDAO trainerDAO) {
-		this.trainerDAO = trainerDAO;
 	}
 
 	@PostConstruct
@@ -97,28 +62,6 @@ public class EmailService {
 		//long interval = TimeUnit.DAYS.toMillis(DAYS_IN_WEEK);
 		long interval = 20000;
 		//timer.scheduleAtFixedRate(this.mailer, startDate, interval);
-
-		this.checkGrades();
-	}
-
-	public void checkGrades() {
-		List<Trainer> trainers = this.trainerDAO.findAll();
-		for (Trainer trainer : trainers) {
-			List<Batch> trainerBatches = this.batchDAO.findAllByTrainer(trainer.getTrainerId());
-			for (Batch batch : trainerBatches) {
-				List<Assessment> batchAssessments = this.assessmentDAO.findByBatchId(batch.getBatchId());
-				List<Trainee> batchTrainees = this.traineeDAO.findAllByBatch(batch.getBatchId());
-				int expectedNumberOfGrades = batchAssessments.size() * batchTrainees.size();
-				int actualNumberOfGrades = 0;
-				for (Assessment assessment : batchAssessments) {
-					List<Grade> assessmentGrades = gradeDAO.findByAssessment(assessment.getAssessmentId());
-					actualNumberOfGrades += assessmentGrades.size();
-				}
-				if (actualNumberOfGrades < expectedNumberOfGrades) {
-					System.out.println("\n" + trainer.getName() + " needs to submit grades" + "\n");
-				}
-			}
-		}
 	}
 
 }
